% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{filter_melt_join}
\alias{filter_melt_join}
\title{melt_join either including or excluding a certain pattern}
\usage{
filter_melt_join(pattern, inv, res, prot_to_bc)
}
\description{
melt_join either including or excluding a certain pattern
}
