% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{filter_multirun}
\alias{filter_multirun}
\title{Filter multirun data}
\usage{
filter_multirun(
  count_list,
  cache_dir = NULL,
  out_dir = NULL,
  save_outputs = TRUE,
  check_cache = TRUE,
  min_n_nz = 3,
  min_frac_nz = 0.5,
  drop_multirun_strains = TRUE,
  id_order = c("strain", "repl", "plate"),
  verbose = TRUE
)
}
\description{
Filter multirun data to ixns with at least min_n_nz nonzero
  output counts or at least a proportion of nonzero counts > min_frac_nz
}
