% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{get_summary}
\alias{get_summary}
\title{Summarise fits}
\usage{
get_summary(
  fit_summaries,
  bead_binders,
  concordance_scores,
  weak_score_threshold = 0.5,
  strong_score_threshold = 1,
  weak_concordance = 0.75,
  strong_concordance = 0.95,
  out_dir,
  verbose = TRUE
)
}
\description{
Summarise fits
}
\details{
this function calls hits from the fit summaries. You can change the
  concordance and score cutoffs, but changing the posterior interval widths
  used to call hits is harder because you'd have to also pass those values to
  the part where the fit summaries are computed.
}
