% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{hit_calling_venn}
\alias{hit_calling_venn}
\title{Venn diagram of hit-calling thresholds}
\usage{
hit_calling_venn(
  score_df,
  score_threshold = 0.5,
  concordance_threshold = 0.75,
  interval_width = 0.95
)
}
\description{
This function draws a Venn diagram of the counts of interactions that pass the three
  hit-calling criteria for the provided thresholds.
}
