% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{melt_join}
\alias{melt_join}
\title{Convert a table of basehit counts from wide to long and join on the header
line giving the mapping from barcodes to proteins}
\usage{
melt_join(count_df, prot_to_bc)
}
\description{
Convert a table of basehit counts from wide to long and join on the header
line giving the mapping from barcodes to proteins
}
