% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{p_type_s}
\alias{p_type_s}
\title{Stats function to calculate the probability of a type S error from a set of
posterior samples}
\usage{
p_type_s(x)
}
\description{
Stats function to calculate the probability of a type S error from a set of
posterior samples
}
