% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{read_multirun}
\alias{read_multirun}
\title{Read in the csvs from count path. Split into three for Pre, beads, and
output.}
\usage{
read_multirun(
  count_path,
  count_threshold = 4,
  cache_dir = NULL,
  out_dir = NULL,
  verbose = TRUE
)
}
\description{
Read in the csvs from count path. Split into three for Pre, beads, and
output.
}
\details{
Drop barcodes with less than or equal to count_threshold counts in
  the pre sample
}
