% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{read_outputs_one}
\alias{read_outputs_one}
\title{Read the count file for one run}
\usage{
read_outputs_one(count_file, prot_to_bc, bh_wr_input)
}
\description{
Read the count file for one run
}
