% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the basehit model}
\usage{
run_model(
  count_path,
  out_dir = "outputs/bh_out/",
  out_name = "results.xlsx",
  model_path = "src/stan_files/basehit_reduce_redundancy_rs_nonsquare.stan"
)
}
\arguments{
\item{out_dir}{path to output, must end with a slash}
}
\description{
Run the basehit model
}
