% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{write_summary}
\alias{write_summary}
\title{Write out fit summary files with notes on bead binding}
\usage{
write_summary(fit_summary, out_dir, verbose = TRUE)
}
\description{
Write out fit summary files with notes on bead binding
}
