% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{hit_calling_venn}
\alias{hit_calling_venn}
\title{Venn diagram of hit-calling thresholds}
\usage{
hit_calling_venn(
  score_df,
  score_threshold = 0.5,
  concordance_threshold = 0.75,
  interval_width = 0.95
)
}
\arguments{
\item{score_df}{a dataframe of basehit scores from model_proteins_separately()}

\item{score_threshold}{exclude ixns with scores below this value}

\item{concordance_threshold}{exclude ixns with concordance values below this threshold}

\item{interval_width}{exclude ixns where a credible interval of this proportion includes 0}
}
\description{
This function draws a Venn diagram of the counts of interactions that pass the three
  hit-calling criteria for the provided thresholds.
}
