% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{model_proteins_separately}
\alias{model_proteins_separately}
\title{Run the basehit model one protein at a time}
\usage{
model_proteins_separately(
  count_path,
  out_dir = "outputs/bh_out/",
  cache_dir = "outputs/bh_cache/",
  split_data_dir = NULL,
  ixn_prior_width = 0.15,
  algorithm = "variational",
  iter_sampling = 5000,
  save_split = TRUE,
  save_fits = FALSE,
  bead_binding_threshold = 1,
  save_bead_binders = TRUE,
  min_n_nz = 3,
  min_frac_nz = 0.5,
  verbose = TRUE,
  seed = 1234
)
}
\arguments{
\item{count_path}{The path to a Mapped_bcs.csv file}

\item{out_dir}{path to output, must end with a slash}

\item{iter_sampling}{number of posterior samples to draw}

\item{bead_binding_threshold}{proteins with enrichment in the beads above this threshold get noted in the output}
}
\description{
Run the basehit model one protein at a time
}
\details{
The count file should have the first row specifying proteins, the
  second specifying barcodes, and all others after that specifying the output
  counts for each strain counts for each barcode (i.e. wide format, strain x
  barcode)

  Implemented with furrr, so run a call to plan() that's appropriate for your
  system in order to parallelize
}
