% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the basehit model}
\usage{
run_model(
  count_path,
  out_dir = "outputs/bh_out/",
  out_name = "results.xlsx",
  model_path = "src/stan_files/basehit_reduce_redundancy_rs_nonsquare.stan",
  vb_threads = 4
)
}
\arguments{
\item{count_path}{The path to a Mapped_bcs.csv file}

\item{out_dir}{path to output, must end with a slash}

\item{out_name}{what you'd like the summary xlsx file to be called}

\item{model_path}{path to the stan model file to use}

\item{vb_threads}{The number of threads to use for the vb fit}
}
\description{
Run the basehit model
}
\details{
The count file should have the first row specifying proteins, the second specifying
  barcodes, and all others after that specifying the output counts for each strain counts for
  each barcode (i.e. wide format, strain x barcode)

  Changing the stan model will almost certainly break things if it has a different
  parameterization as the default.
}
