% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_proteins_separately.R
\name{write_summary}
\alias{write_summary}
\title{Write out fit summary files with notes on bead binding}
\usage{
write_summary(
  fit_summaries,
  bead_binders,
  concordance_scores,
  weak_score_threshold = 0.5,
  strong_score_threshold = 1,
  weak_concordance = 0.75,
  strong_concordance = 0.95,
  verbose = TRUE
)
}
\description{
Write out fit summary files with notes on bead binding
}
